requirements.txt
# VDS Jira Ticket Assistant — Requirements
# Compatible with Python 3.9+

# ── Streamlit UI ────────────────────────────────────────
streamlit==1.35.0

# ── AI / Embeddings ─────────────────────────────────────
anthropic==0.29.0
sentence-transformers==3.0.1

# ── Vector Search ────────────────────────────────────────
faiss-cpu==1.8.0

# ── OCR ──────────────────────────────────────────────────
pytesseract==0.3.10
Pillow==10.4.0

# ── Data ─────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Scheduler ────────────────────────────────────────────
APScheduler==3.10.4

# ── Environment variables ─────────────────────────────────
python-dotenv==1.0.1

# ── HTTP (for Anthropic SDK) ─────────────────────────────
httpx==0.27.0
